+++ sound-mind-studio/tests/main.cpp
@@ -23,6 +23,7 @@
 #include "test_layers_panel.h"
 #include "test_loop_panel.h"
 #include "test_main_window.h"
+#include "test_paint_controller.h"
 #include "test_playback_controller.h"
 #include "test_playback_panel.h"
 #include "test_record_panel.h"
@@ -78,6 +79,9 @@ int main(int argc, char** argv) {
 
     PlaybackControllerTest playbackControllerTest;
     status |= QTest::qExec(&playbackControllerTest, argc, argv);
+
+    PaintControllerTest paintControllerTest;
+    status |= QTest::qExec(&paintControllerTest, argc, argv);
 
     AudioSnippetPickerDialogTest audioSnippetPickerDialogTest;
     status |= QTest::qExec(&audioSnippetPickerDialogTest, argc, argv);